move CTransaction to primitives
