+++ src/Makefile.am
@@ -127,6 +127,8 @@ GRIDCOIN_CORE_H = \
     gridcoin/voting/result.h \
     gridcoin/voting/vote.h \
     hash.h \
+    index/disktxpos.h \
+    index/txindex.h \
     init.h \
     key.h \
     keystore.h \
@@ -137,7 +139,9 @@ GRIDCOIN_CORE_H = \
     netbase.h \
     net.h \
     pbkdf2.h \
+    policy/fees.h \
     prevector.h \
+    primitives/transaction.h \
     protocol.h \
     reverselock.h \
     rpc/blockchain.h \
@@ -167,6 +171,7 @@ GRIDCOIN_CORE_H = \
     util/threadnames.h \
     util/time.h \
     util.h \
+    validation.h \
     version.h \
     wallet/coincontrol.h \
     wallet/db.h \
@@ -207,6 +212,7 @@ GRIDCOIN_CORE_CPP = addrdb.cpp \
     gridcoin/superblock.cpp \
     gridcoin/support/block_finder.cpp \
     gridcoin/tally.cpp \
+    gridcoin/tx_message.cpp \
     gridcoin/upgrade.cpp \
     gridcoin/voting/builders.cpp \
     gridcoin/voting/claims.cpp \
@@ -224,6 +230,8 @@ GRIDCOIN_CORE_CPP = addrdb.cpp \
     net.cpp \
     noui.cpp \
     pbkdf2.cpp \
+    policy/fees.cpp \
+    primitives/transaction.cpp \
     protocol.cpp \
     rpc/blockchain.cpp \
     rpc/client.cpp \
@@ -249,6 +257,7 @@ GRIDCOIN_CORE_CPP = addrdb.cpp \
     util/threadnames.cpp \
     util/time.cpp \
     util.cpp \
+    validation.cpp \
     version.cpp \
     wallet/db.cpp \
     wallet/rpcdump.cpp \
