+++ src/policy/fees.cpp
@@ -1,0 +1,50 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2020 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include "consensus/consensus.h"
+#include "policy/fees.h"
+
+CAmount GetBaseFee(const CTransaction& tx, enum GetMinFee_mode mode)
+{
+    // Base fee is either MIN_TX_FEE or MIN_RELAY_TX_FEE
+    CAmount nBaseFee = (mode == GMF_RELAY) ? MIN_RELAY_TX_FEE : MIN_TX_FEE;
+
+    // For block version 11 onwards, which corresponds to CTransaction::CURRENT_VERSION 2,
+    // a multiplier is used on top of MIN_TX_FEE and MIN_RELAY_TX_FEE
+    if (tx.nVersion >= 2)
+    {
+        nBaseFee *= 10;
+    }
+
+    return nBaseFee;
+}
+
+CAmount GetMinFee(const CTransaction& tx, unsigned int nBlockSize, enum GetMinFee_mode mode, unsigned int nBytes)
+{
+    CAmount nBaseFee = GetBaseFee(tx, mode);
+
+    unsigned int nNewBlockSize = nBlockSize + nBytes;
+    CAmount nMinFee = (1 + (int64_t)nBytes / 1000) * nBaseFee;
+
+    // To limit dust spam, require MIN_TX_FEE/MIN_RELAY_TX_FEE if any output is less than 0.01
+    if (nMinFee < nBaseFee)
+    {
+        for (auto const& txout : tx.vout)
+            if (txout.nValue < CENT)
+                nMinFee = nBaseFee;
+    }
+
+    // Raise the price as the block approaches full
+    if (nBlockSize != 1 && nNewBlockSize >= MAX_BLOCK_SIZE_GEN/2)
+    {
+        if (nNewBlockSize >= MAX_BLOCK_SIZE_GEN)
+            return MAX_MONEY;
+        nMinFee *= MAX_BLOCK_SIZE_GEN / (MAX_BLOCK_SIZE_GEN - nNewBlockSize);
+    }
+
+    if (!MoneyRange(nMinFee))
+        nMinFee = MAX_MONEY;
+    return nMinFee;
+}
