+++ src/qt/coincontroldialog.cpp
@@ -6,6 +6,8 @@
 #include "walletmodel.h"
 #include "addresstablemodel.h"
 #include "optionsmodel.h"
+#include "policy/fees.h"
+#include "validation.h"
 #include "wallet/coincontrol.h"
 
 #include <QApplication>
@@ -488,7 +490,7 @@ void CoinControlDialog::updateLabels(WalletModel *model, QDialog* dialog)
         int64_t nFee = nTransactionFee * (1 + (int64_t)nBytes / 1000);
 
         // Min Fee
-        int64_t nMinFee = txDummy.GetMinFee(1000, GMF_SEND, nBytes);
+        int64_t nMinFee = GetMinFee(txDummy, 1000, GMF_SEND, nBytes);
 
         nPayFee = max(nFee, nMinFee);
