+++ src/wallet/walletdb.cpp
@@ -219,7 +219,7 @@ ReadKeyValue(CWallet* pwallet, CDataStream& ssKey, CDataStream& ssValue,
             ssKey >> hash;
             CWalletTx& wtx = pwallet->mapWallet[hash];
             ssValue >> wtx;
-            if (wtx.CheckTransaction() && (wtx.GetHash() == hash))
+            if (CheckTransaction(wtx) && (wtx.GetHash() == hash))
                 wtx.BindWallet(pwallet);
             else
             {
